+++ Localization/zhCN.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-06-23 17-58-06 ]] --
+-- [[ Exported at 2022-06-25 21-14-11 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --