+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 9.2.5.36.0 - 2022-
+### Removed
+- Achievement covenant information
+- Covenant filters
+
 ## 9.2.5.35.0 - 2022-06-25
 ### Added
 - Calendar view for earned achievements (only 1st 4 per day for now)